refactor: change package and add controller
