+++ src/main/java/step3/domain/Position.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 public class Position {
     private int position;
