+++ src/main/java/step3/domain/Score.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.domain;
+
+import step3.domain.Car;
 
 import java.util.ArrayList;
 import java.util.Collections;
