+++ src/main/java/step3/service/CarEngine.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.service;
 
 @FunctionalInterface
 public interface CarEngine {
