+++ src/main/java/step3/service/RacingCarGame.java
@@ -1,4 +1,9 @@
-package step3;
+package step3.service;
+
+import step3.domain.Car;
+import step3.domain.Cars;
+import step3.view.InputView;
+import step3.view.PrintView;
 
 import java.util.ArrayList;
 import java.util.List;
