+++ src/main/java/step3/view/InputValidator.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.view;
 
 public class InputValidator {
