+++ src/test/java/step3/NameTest.java
@@ -3,6 +3,7 @@ package step3;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Name;
 
 public class NameTest {
