+++ src/test/java/step3/ScoreTest.java
@@ -3,6 +3,10 @@ package step3;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Car;
+import step3.domain.Cars;
+import step3.domain.Score;
+import step3.service.CarEngine;
 
 import java.util.ArrayList;
 import java.util.List;
