+++ src/test/java/step3/TestEngine.java
@@ -1,5 +1,7 @@
 package step3;
 
+import step3.service.CarEngine;
+
 public class TestEngine implements CarEngine {
     private static final int FORWARD = 4;
     private final int power;
